+++ src/worker/runthis.rs
@@ -18,7 +18,7 @@ pub fn run(mut metric: Metric) -> Metric {
             metric.value = Some(value as f64);
             metric.units = None;
             metric.message = Some("OK".to_string());
-            metric.graph_value = Some(value as u32);
+            metric.graph_value = Some(value as i64);
             metric.graph_short_name = Some(metric.short_name.clone());
 
         },
@@ -27,7 +27,7 @@ pub fn run(mut metric: Metric) -> Metric {
             metric.value = Some(-1.0);
             metric.units = None;
             metric.message = Some("Failed to execute command".to_string());
-            metric.graph_value = Some(0 as u32);
+            metric.graph_value = Some(0 as i64);
             metric.graph_short_name = Some(metric.short_name.clone());
         },
     }
